refactor(log): extracted log preview component to separate lib
This is needed since this log preview component will be used in
multiple places in application
