+++ libs/bublik/features/log-preview-drawer/README.md
@@ -1,0 +1,7 @@
+# log-preview-drawer
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test log-preview-drawer` to execute the unit tests via [Vitest](https://vitest.dev/).
