+++ libs/bublik/features/log-preview-drawer/project.json
@@ -1,0 +1,20 @@
+{
+	"name": "log-preview-drawer",
+	"$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+	"sourceRoot": "libs/bublik/features/log-preview-drawer/src",
+	"projectType": "library",
+	"tags": [],
+	"targets": {
+		"lint": {
+			"executor": "@nx/eslint:lint"
+		},
+		"test": {
+			"executor": "@nx/vite:test",
+			"outputs": ["{options.reportsDirectory}"],
+			"options": {
+				"reportsDirectory": "../../../../coverage/libs/bublik/features/log-preview-drawer",
+				"passWithNoTests": true
+			}
+		}
+	}
+}
