+++ libs/bublik/features/log-preview-drawer/src/index.ts
@@ -1,0 +1,3 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+/* SPDX-FileCopyrightText: 2024 OKTET LTD */
+export * from './log-preview-drawer.container';
