+++ libs/bublik/features/log-preview-drawer/vite.config.ts
@@ -1,0 +1,29 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
+
+export default defineConfig({
+	root: __dirname,
+	cacheDir:
+		'../../../../node_modules/.vite/libs/bublik/features/log-preview-drawer',
+
+	plugins: [react(), nxViteTsPaths()],
+
+	// Uncomment this if you are using workers.
+	// worker: {
+	//  plugins: [ nxViteTsPaths() ],
+	// },
+
+	test: {
+		watch: false,
+		globals: true,
+		environment: 'jsdom',
+		include: ['src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
+		reporters: ['default'],
+		coverage: {
+			reportsDirectory:
+				'../../../../coverage/libs/bublik/features/log-preview-drawer',
+			provider: 'v8'
+		}
+	}
+});
