+++ package.json
@@ -156,6 +156,7 @@
 		"@svgr/webpack": "8.1.0",
 		"@swc-node/register": "1.9.2",
 		"@swc/core": "1.5.7",
+		"@swc/helpers": "~0.5.11",
 		"@tailwindcss/forms": "^0.5.3",
 		"@testing-library/dom": "^9.3.1",
 		"@testing-library/jest-dom": "^5.17.0",
