+++ tsconfig.base.json
@@ -83,7 +83,10 @@
 			"@/shared/hooks": ["libs/shared/hooks/src/index.ts"],
 			"@/shared/tailwind-ui": ["libs/shared/tailwind-ui/src/index.ts"],
 			"@/shared/types": ["libs/shared/types/src/index.ts"],
-			"@/shared/utils": ["libs/shared/utils/src/index.ts"]
+			"@/shared/utils": ["libs/shared/utils/src/index.ts"],
+			"@/bublik/features/log-preview-drawer": [
+				"libs/bublik/features/log-preview-drawer/src/index.ts"
+			]
 		}
 	},
 	"exclude": ["node_modules", "tmp"]
