+++ vitest.workspace.ts
@@ -1,0 +1,1 @@
+export default ['**/*/vite.config.ts', '**/*/vitest.config.ts'];
